Update XSS.py
Rough version that attempts to directly access a form with known fields on test website. Still doesn't work fully, as it's not able to submit the button completely (Probably failing a validation check somewhere, not quite sure). This is largely in RoboBrowser to experiment with it and see what can be done (Despite the lack of documentation), so this code will be subject to changes to bring it in line with other modules down the line.

Also, the code above may not run if the whole thing is copied, just remove Class(xss): and it should work for local testing.