Update XSS.py
Rewrote XSS to be more in line with what's currently there for SQL Injection for Reflected XSS. The Stored XSS is currently disabled, but that will be looked at after the Reflected is completed. Currently trying to figure out how to determine if the injected script executed successfully or not.