Create XSSAttacks.txt
There's only 2 attacks at the moment (One that runs <script> and one that triggers a <body> statement. Will add more as we go along, but we only really need to detect the presence of the < and > characters not being sanitized properly. Need to look at other possible XSS scripts to test, but these two are enough to detect the presence of XSS.