Update XSS.py
Updated it to use a file to check for sample-scripts to attempt to inject, much like the SQL code. 

There's no real difference between attackStored and attackReflect, largely due to how inputs are obtained by link.getInputs(). All the code really does is just find fields, attempt to do a POST, and see if the script ends up in the resulting page somewhere. This does have some issues though, as the POST doesn't seem to change to new pages if that occurs.