Update main.py
Added rough version of XSS, not sure if the commented out section with classification.vulnerability works yet, but the rest should be fine, which is why it's being pushed. It's also using the Fuzz methods, so I updated import to include Fuzz.py (SQL should use that as well based off the current test.py file), along with the XSS.py file.